+++ doc/release-notes.md
@@ -13,3 +13,6 @@ This release includes the following features and fixes:
    to relay to miners).
  - Added `nTx` return value to `getblock` and `getblockheader` detailing the
    number of transactions in the returned block.
+ - Remove deprecated "startingpriority" and "currentpriority" from
+   `getrawmempool`, `getmempoolancestors`, `getmempooldescendants` and
+   `getmempoolentry` RPC.